add brillig_call submodule
