+++ compiler/noirc_evaluator/src/brillig/brillig_gen.rs
@@ -1,11 +1,10 @@
 //! The code generation logic for converting [crate::ssa] objects into their respective [Brillig] artifacts.  
-pub(crate) mod brillig_black_box;
 pub(crate) mod brillig_block;
 pub(crate) mod brillig_block_variables;
+mod brillig_call;
 pub(crate) mod brillig_fn;
 pub(crate) mod brillig_globals;
 mod brillig_instructions;
-pub(crate) mod brillig_slice_ops;
 pub(crate) mod constant_allocation;
 mod variable_liveness;
