+++ compiler/noirc_evaluator/src/brillig/brillig_gen/brillig_call/brillig_slice_ops.rs
@@ -7,7 +7,7 @@ use crate::brillig::brillig_ir::{
     registers::RegisterAllocator,
 };
 
-use super::brillig_block::BrilligBlock;
+use super::super::brillig_block::BrilligBlock;
 
 impl<Registers: RegisterAllocator> BrilligBlock<'_, Registers> {
     fn write_variables(&mut self, write_pointer: MemoryAddress, variables: &[BrilligVariable]) {
