+++ compiler/noirc_evaluator/src/brillig/brillig_gen/brillig_instructions/mod.rs
@@ -1,3 +1,2 @@
 pub(crate) mod brillig_binary;
-pub(crate) mod brillig_calls;
 pub(crate) mod brillig_memory;
